Break up AlongStepImpl into one file per functor
